==> Code/R/02_Create_Test_Features_Master_sparse.R <==
rm(list=ls())
library(data.table)
set.seed(12345)

source("02_Create_Test_Features_sparse.R")
create_test_features("GA", "FLtrain_NCtest_GAtest/FL_Train_s.csv", "FLtrain_NCtest_GAtest/GAtest_FLtrain_s.csv")
create_test_features("NC", "FLtrain_NCtest_GAtest/FL_Train_s.csv", "FLtrain_NCtest_GAtest/NCtest_FLtrain_s.csv")
create_test_features("FL", "GAtrain_NCtest_FLtest/GA_Train_s.csv", "GAtrain_NCtest_FLtest/FLtest_GAtrain_s.csv")
create_test_features("NC", "GAtrain_NCtest_FLtest/GA_Train_s.csv", "GAtrain_NCtest_FLtest/NCtest_GAtrain_s.csv")
create_test_features("FL", "NCtrain_FLtest_GAtest/NC_Train_s.csv", "NCtrain_FLtest_GAtest/FLtest_NCtrain_s.csv")
create_test_features("GA", "NCtrain_FLtest_GAtest/NC_Train_s.csv", "NCtrain_FLtest_GAtest/GAtest_NCtrain_s.csv")
create_test_features("NC_white", "NCtrain_NCtest_white/NC_rest_Train_s.csv", "NCtrain_NCtest_white/NCwtest_NCtrain_s.csv")
create_test_features("NC_white", "FLGAtrain_NCtest/FLGA_Train_s.csv", "NCtrain_NCtest_white/NCwtest_FLGAtrain_s.csv")
create_test_features("NC", "FLGAtrain_NCtest/FLGA_Train_s.csv", "FLGAtrain_NCtest/NCtest_FLGAtrain_s.csv")
create_test_features("CaseStudy", "FLGANCtrain_ContrPred/FLGANC_Train_s.csv", "FLGANCtrain_ContrPred/ContTest_FLGANCtrain_s.csv")
create_test_features("CaseStudyGA", "FLGANCtrain_ContrPred/FLGANC_Train_s.csv", "FLGANCtrain_ContrPred/ContTest_GA_FLGANCtrain_s.csv")
